requirements.txt
streamlit>=1.32.0
openai>=1.30.0
ollama>=0.4.0
qdrant-client>=1.9.0
pypdf>=4.2.0
pypdfium2>=4.30.0
pandas>=2.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3
tiktoken>=0.7.0
pydantic>=2.7.0
numpy>=1.26.0
sentence-transformers>=2.7.0
torch>=2.2.0
pytesseract>=0.3.10
Pillow>=10.3.0
docx2python>=3.5.0
beautifulsoup4>=4.12.3
openpyxl>=3.1.2
